Move betty.service.plugin to betty.plugin.service
