+++ betty/app/__init__.py
@@ -27,11 +27,11 @@ from betty.locale.translation import (
     TranslationRepository,
 )
 from betty.multiprocessing import ProcessPoolExecutor
+from betty.plugin.service import ServicePluginProvider, ServicePlugins, SupportPlugins
 from betty.portable.file import assert_load_file
 from betty.service.factory import DataManufacturable
 from betty.service.level import ChainedServiceLevel, Plugins, ServiceLevel
 from betty.service.level.universe import UNIVERSE
-from betty.service.plugin import ServicePluginProvider, ServicePlugins, SupportPlugins
 from betty.service.provider import ServiceFactory, service
 from betty.typing import threadsafe
 from betty.user.no_op import NoOpUser
