+++ betty/jinja/test.py
@@ -10,7 +10,7 @@ from typing import final, override
 from betty.locale.localizable.gettext import _, ngettext
 from betty.plugin import Plugin, PluginTypeDefinition
 from betty.plugin.factory import PluginManufacturer
-from betty.service.plugin import ServicePluginDefinition
+from betty.plugin.service import ServicePluginDefinition
 
 
 class JinjaTest(Plugin["JinjaTestDefinition"], ABC):
