+++ betty/link.py
@@ -11,7 +11,7 @@ from betty.machine_name import ResolvableMachineName
 from betty.plugin import Plugin, PluginTypeDefinition
 from betty.plugin.factory import PluginManufacturer
 from betty.plugin.ordered import Order, OrderedPluginDefinition
-from betty.service.plugin import ServicePluginDefinition
+from betty.plugin.service import ServicePluginDefinition
 
 
 class LinkType(Protocol):
